+++ scripts/update-assembler.js
@@ -6,7 +6,7 @@
  * @todo Publish the assembler to NPM to always keep it up to date.
  */
 'use strict';
-const U = require('./utilities.js');
+const U = require('../app/utilities.js');
 const files = U.getFilesInFolder('../../highcharts/', 'assembler', true);
 files.forEach(file => {
 	U.copyFile('../../highcharts/' + file, '../' + file);